[Src] Promote vkEnumeratePhysicalDeviceGroupsKHX command; remove experimental
